+++ Bubbles/bubble.h
@@ -9,11 +9,11 @@ class bubble : public collidable
 public:
     // Public variables
     point3 center;
-    float radius;
+    double radius;
 
 public:
     // Public functions
     bubble(point3 _center, double _radius);
-    virtual bool hit(const ray& _ray, float t_min, float t_max, objectRecord& rec) const override;
+    virtual bool hit(const ray& _ray, double t_min, double t_max, objectRecord& rec) const override;
 };
